move all the hamming-distance functions around
