+++ hamdis/hamming-distance.js
