+++ hamdis/variations/hamdis-original.r
