+++ hamdis/variations/hamdis.r
