+++ hamdis/variations/hamming-distance-2.cpp
