+++ hamdis/variations/hamming-distance-2.js
